requirements.txt
langchain==0.1.1
langchain_community==0.0.13
langchain_experimental==0.0.49
openai==1.9.0
Pillow==10.2.0
protobuf==4.25.2
PyPDF2==3.0.1
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.28.2
streamlit_chat==0.1.1
google-cloud-storage==2.14.0
pymysql